--- a/mentat/embeddings.py
+++ b/mentat/embeddings.py
@@ -105,13 +105,8 @@ async def get_feature_similarity_scores(
         raise MentatError(f"Missing model context size for {EMBEDDING_MODEL}.")
 
     # Keep things in the same order
-    t1 = default_timer()
     checksums: list[str] = [f.get_checksum() for f in features]
-    t2 = default_timer()
-    print('Got checksums in', t2 - t1)
     tokens: list[int] = await count_feature_tokens(features, EMBEDDING_MODEL)
-    t3 = default_timer()
-    print('Got tokens in', t3 - t2)
 
     # Make a checksum:content dict of all items that need to be embedded
     items_to_embed = dict[str, str]()
@@ -130,8 +125,6 @@ async def get_feature_similarity_scores(
             items_to_embed[checksum] = "\n".join(feature_content)
             items_to_embed_tokens[checksum] = token
             num_prompt_tokens += token
-    t4 = default_timer()
-    print('Got items to embed in', t4 - t3)
 
     # If it costs more than $1, get confirmation from user.
     cost = model_price_per_1000_tokens(EMBEDDING_MODEL)
@@ -153,23 +146,13 @@ async def get_feature_similarity_scores(
 
     # Fetch embeddings in batches
     batches = _batch_ffd(items_to_embed_tokens, max_model_tokens)
-    t5 = default_timer()
-    print('Got batches in', t5 - t4)
 
     _start_time = default_timer()
-    _embed_time = 0.
-    _add_to_db_time = 0.
     for i, batch in enumerate(batches):
         batch_content = [items_to_embed[k] for k in batch]
         stream.send(f"Embedding batch {i + 1}/{len(batches)}...")
-        t1a = default_timer()
         response = await call_embedding_api(batch_content, EMBEDDING_MODEL)
-        t1b = default_timer()
-        print('Got response in', t1b - t1a)
-        _embed_time += t1b - t1a
         database.set({k: v for k, v in zip(batch, response)})
-        t1c = default_timer()
-        _add_to_db_time += t1c - t1b
     if len(batches) > 0:
         cost_tracker.display_api_call_stats(
             num_prompt_tokens,
@@ -178,16 +161,10 @@ async def get_feature_similarity_scores(
             default_timer() - _start_time,
             decimal_places=4,
         )
-    print(f'Total time to embed {len(batches)} batches:', _embed_time)
-    print('Total time to add to db:', _add_to_db_time)
 
     # Calculate similarity score for each feature
     prompt_embedding = database.get([prompt_checksum])[prompt_checksum]
-    t6 = default_timer()
     embeddings = database.get(checksums)
-    t7 = default_timer()
-    print('Total time to get from db:', t7 - t6)
     scores = [_cosine_similarity(prompt_embedding, embeddings[k]) for k in checksums]
-    print('Total time to cosine:', default_timer() - t7)
 
     return scores