Two changes:
- Use sqlite database instead of json/gzip. We save ~10s by not writing the gzip file, and lose ~0.3 seconds reading from the database instead of just from memory.
- Make up to n=10 (arbitrarily) simultaneous calls to the embeddings API, handle minute-level RateLimits

A future optimization could be to use [Annoy](https://github.com/spotify/annoy). This would replace that 0.3s call to the database and ~0.1s calculating cosine similarity. 

(all these numbers are from running mentat on the mentat repo)